chore: update CI configuration and enhance build process
- Set Node.js version to 22 and enforced npm ci for consistent dependency installation.
- Added environment variable to force JavaScript actions to Node 24.
- Included a step to run frontend unit tests in the CI pipeline.
- Updated build script to embed Windows .exe icon and improved config version handling in AppConfig.
- Enhanced transcription logic to handle empty results and detect prompt echoes in STT processing.

+++ src-tauri/src/stt/sensevoice.rs
@@ -99,7 +99,7 @@ fn transcribe_via_websocket(audio: &[f32], sample_rate: u32) -> anyhow::Result<S
     let samples_bytes = (audio.len() * 4) as u32;
     payload.extend_from_slice(&samples_bytes.to_le_bytes());
     for &s in audio {
-        payload.extend_from_slice(&(s as f32).to_le_bytes());
+        payload.extend_from_slice(&s.to_le_bytes());
     }
 
     let rt = tokio::runtime::Builder::new_current_thread()
